Fixed errors while plc is not connected / moved inteface.py
